backend/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asttokens==2.4.0
backcall==0.2.0
beautifulsoup4==4.12.3
biopython==1.85
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
dnspython==2.8.0
email-validator==2.3.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.0
fastapi==0.109.0
Flask==3.0.3
fonttools==4.55.3
greenlet==3.3.0
h11==0.16.0
httptools==0.7.1
idna==3.6
ipykernel==6.25.2
ipython==8.16.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.4.0
jupyter_core==5.4.0
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.6
mysql-connector-python==8.0.28
nest-asyncio==1.5.8
numpy==1.26.1
opencv-contrib-python==4.9.0.80
opencv-python==4.8.1.78
openpyxl==3.1.2
packaging==23.2
pandas==2.2.2
parso==0.8.3
pickleshare==0.7.5
pillow==10.2.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
protobuf==3.19.4
psutil==5.9.5
psycopg2-binary==2.9.11
pure-eval==0.2.2
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
Pygments==2.16.1
pyparsing==3.2.0
python-dateutil==2.8.2
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2024.1
pywin32==306
PyYAML==6.0.3
pyzbar==0.1.9
pyzmq==25.1.1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.23
stack-data==0.6.3
starlette==0.35.1
tornado==6.3.3
traitlets==5.11.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.27.0
watchfiles==1.1.1
wcwidth==0.2.8
websockets==15.0.1
Werkzeug==3.0.4
